Fix root folder feature
The previous root folder feature was implemented by modifying generated
HTML output. This caused problems with generated RSS files.

This commit replaces the HTML output modifying version with one that
modifies the structure of the whole site S-expression in a structured
manner. The HTML is read in Hiccup format instead of raw HTML text,
which is later manipulated. Besides the content, also the URIs in the
site configuration are also modified to include the custom root.

Additional changes:

- Refactored augmentations to their own modules and added tests.
- Added custom tags for managing root path in Selmer templates.
- Assets are aware of custom root paths.
- Fixed page URNs to not include invalid characters such as slashes.
- URI module is more sophisticated in handling URIs
- Replaced author first & last names with just a name.